v0.2: Python reference reproduces stored gate frontier within ±0.001
Three fidelity fixes:

1. agreement_count: numpy 2.x makes bool+bool=bool (logical OR). The
   training pipeline's ``((h1==h0)+(h2==h0)+(h3==h0)).astype(np.float32)``
   silently produced a 0/1 indicator, not a 0-3 count. Port now matches
   that behavior (cast after sum, not per-term) for artifact compatibility.
   Proper 0-3 count retraining is on the v0.3 roadmap.

2. Boundary-token sets: extract_all_features now accepts precomputed
   ``period_ids`` / ``newline_ids`` kwargs. scripts/reproduce.py scans the
   tokenizer vocab to build them (matches training exactly).

3. Cluster K-means centers: extract_all_features accepts ``cluster_centers``
   (already supported). scripts/reproduce.py now fits K=32 centers on the
   first 20 training sequences with 20 Lloyd iterations, matching
   test_physics_apertures.py::aperture_cluster.

Measured reproduction (skip at each λ):
  λ=0.85  reproduced 0.1913  stored 0.1913
  λ=0.90  reproduced 0.1411  stored 0.1411
  λ=0.95  reproduced 0.0981  stored 0.0989
  λ=0.99  reproduced 0.0602  stored 0.0602

README caveat replaced with reproduction table.

+++ unified_gate/features/base.py
@@ -110,11 +110,18 @@ def build_base_features(
         head_preds.append(head_preds[0])
         head_confs.append(head_confs[0])
 
+    # NOTE: the training pipeline computed this as ``((h1==h0)+(h2==h0)+(h3==h0))
+    # .astype(np.float32)`` — under numpy 2.x bool+bool stays bool (logical OR),
+    # so this feature is secretly a 0/1 "at least one head agrees" indicator,
+    # not a 0-3 count. We reproduce that behavior here to stay faithful to the
+    # trained gate's feature distribution. A corrected retraining lives on the
+    # v0.3 roadmap (feature name stays "agreement_count" for artifact
+    # compatibility).
     agreement = (
-        (head_preds[1] == head_preds[0]).astype(np.float32)
-        + (head_preds[2] == head_preds[0]).astype(np.float32)
-        + (head_preds[3] == head_preds[0]).astype(np.float32)
-    )
+        (head_preds[1] == head_preds[0])
+        + (head_preds[2] == head_preds[0])
+        + (head_preds[3] == head_preds[0])
+    ).astype(np.float32)
     conf_stack = np.stack(head_confs[:4], axis=1)
     conf_var = conf_stack.var(axis=1)
     conf_min = conf_stack.min(axis=1)
